+++ src/command/DevCount.php
@@ -14,7 +14,6 @@
  */
 namespace Rocareer\WebmanDev\command;
 
-use support\think\Db;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;